Move ws_statement type into new file
