+++ src/command/command.h
@@ -32,40 +32,9 @@
 #include <stdbool.h>
 
 // forward declarations
+struct ws_command_args;
 struct ws_processor;
 union ws_value_union;
-
-/**
- * A command argument
- *
- * There are two types of arguments: the ones passed directly and the ones which
- * are already on the stack.
- */
-struct ws_argument {
-    enum {
-        indirect, //!< argument is given by position
-        direct //!< argument is given directly
-    } type; //!< @public type of argument (direct or indirect)
-    union {
-        ssize_t pos; //!< position, negative ones are from the top of the stack
-        struct ws_value* val; //!< a value
-    } arg; //!< @public the argument
-}; 
-
-/**
- * Command arguments
- * 
- * This struct represents command arguments applied to a command as part of a
- * statement.
- *
- * @note If `vals` is `NULL`, it indicates that the `num` topmost elements which
- *       are currently on the stack are arguments which should be passed to the
- *       command.
- */
-struct ws_command_args {
-    size_t num; //!< @public number of arguments
-    struct ws_argument* vals; //!< @public arguments
-};
 
 /**
  * Function type for command implementation
@@ -105,17 +74,6 @@ struct ws_command {
 };
 
 /**
- * Statement
- *
- * This datatype represents a statement, which is part of an transaction
- */
-struct ws_statement {
-    struct ws_command const* command; //!< @public command to invoke
-    struct ws_command_args args; //!< @public arguments to invoke the command with
-};
-
-
-/**
  * Find a command by name
  *
  * @returns command with the name given or NULL
@@ -123,16 +81,6 @@ struct ws_statement {
 struct ws_command const*
 ws_command_get(
     char const* name //!< name of the command
-);
-
-/**
- * Deinitialize a statement
- *
- * @return true if deinitialization was successfull, else false
- */
-bool
-ws_statement_deinit(
-    struct ws_statement* self
 );
 
 #endif // __WS_COMMAND_COMMAND_H__
