+++ src/command/statement.c
@@ -1,0 +1,46 @@
+/*
+ * waysome - wayland based window manager
+ *
+ * Copyright in alphabetical order:
+ *
+ * Copyright (C) 2014-2015 Julian Ganz
+ * Copyright (C) 2014-2015 Manuel Messner
+ * Copyright (C) 2014-2015 Marcel Müller
+ * Copyright (C) 2014-2015 Matthias Beyer
+ * Copyright (C) 2014-2015 Nadja Sommerfeld
+ *
+ * This file is part of waysome.
+ *
+ * waysome is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation, either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * waysome is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with waysome. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "command/statement.h"
+#include "values/value.h"
+
+bool
+ws_statement_deinit(
+    struct ws_statement* self
+) {
+    while (self->args.num--) {
+        if (self->args.vals[self->args.num].type == direct) {
+            ws_value_deinit(self->args.vals[self->args.num].arg.val);
+            free(self->args.vals[self->args.num].arg.val);
+        }
+    }
+
+    free(self->args.vals);
+
+    return true;
+}
+
